Add files via upload
# p-comp-week-1-labs-leahoho
p-comp-week-1-labs-leahoho created by GitHub Classroom
---
We are in a world where robots coexist

Due to the development of technology, it is very difficult for us to distinguish between humans and robots in appearance, but there is a big difference in the thinking systems. By default, robots can only think rationally without adding human emotions.  However, robots have become patient with humans and slowly learn human emotions, but they have not been fully used. Therefore, it is difficult for humans to get along with robots, which makes it difficult to communicate.

This device allows us to distinguish what mode the robot uses when getting along with humans，and know whether the robot adds emotions when communicating.  When robots express rational opinions, the led will show blue color. On the contrary, they can press the green button when they express their opinions with emotion. This can help us to have better communication with robots and to distinguish whether the thinking mode is rational or emotional.