Add files via upload
Reflections from Building the Hand:

We are habitual to believing in our feelings and controlled by the body, so when I use the robotic hand in class, I cannot wholly trust this hand. Because it is difficult to control and unable to connect with our brains, I feel frustrated and overwhelmed. After all, we don't have any emotional connection with A. I or technology.

After seeing art performances of Viktoria Modesta, she used 3-D printed nails and sensors on her wrists to trigger sound effects and shifts in lighting on her skin. This project inspired me to think can we feel the connection or think it is a part of our body when we install different machines on our body or are we treating it as Tools. And how can we accept transforming our bodies with technologies?